(refactor) split partial aggregatefn helpers, documentation, and review comments
Similar to the scalar split, but for helpers in `BloomPartial` for `AggreateFnVTable`.
It also addresses review comments around documentation and types.
